requirements.txt
celery==5.2.7
Django==4.1.3
djangorestframework==3.14.0
drf-spectacular==0.25.1
pytest-cov==4.0.0
pytest-django==4.5.2
pytz==2022.6
pyyaml==6.0
redis==4.4.0
requests==2.28.1